+++ forum/settings.py
@@ -40,8 +40,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'authentication',
-    'activity',
-    'category',
     'profiles',
 ]
